+++ test/index_test.js
@@ -166,6 +166,13 @@ describe("FetchBody", function() {
       ;(yield res).body.must.equal(JSON.stringify({key: "value"}))
     })
 
+    it("must set body if Content-Length 0", function*() {
+      var res = fetch(URL, {method: "HEAD"})
+      var headers = {"Content-Type": "text/plain", "Content-Length": "0"}
+      this.requests[0].respond(200, headers, "")
+      ;(yield res).body.must.equal("")
+    })
+
     it("must set body when response not OK", function*() {
       var res = fetch(URL)
       var headers = {"Content-Type": "text/plain"}
@@ -241,6 +248,18 @@ describe("FetchBody", function() {
       res = yield res
       res.must.not.have.property("body")
       yield res.text().must.then.equal(JSON.stringify({key: "value"}))
+    })
+
+    // Some implementations respond to HEAD requests with a Content-Type, but
+    // Content-Lenght of 0. This saves throwing a parse error.
+    it("must not set body if Content-Length 0", function*() {
+      var res = fetch(URL, {method: "HEAD"})
+      var headers = {"Content-Type": "application/json", "Content-Length": "0"}
+      this.requests[0].respond(200, headers, "")
+
+      res = yield res
+      res.must.not.have.property("body")
+      yield res.text().must.then.equal("")
     })
 
     // This was a released bug with Remote that I noticed on Nov 25, 2014 and